px to vh & vw, settings refactor
~80%, still need to change some px, fix some components & redesign parts
