+++ app-front/Components/Button/index.tsx
@@ -1,12 +1,13 @@
 import styled from 'styled-components';
 
 const Button = styled.div`
-    border: 0px;
+    border: 0;
     cursor: pointer;
-    border-radius: 18px;
-    padding: 10px;
-    font-size: 17px;
+    border-radius: var(--small-radius);
+    padding: calc(var(--decade));
+    font-size: var(--decade);
     width: 100%;
+    height: calc(var(--decade) * 4);
     box-sizing: border-box;
     text-align: center;
     background-color: rgba(255, 255, 255, 0.1);
