+++ app-front/Components/FriendsBar/chat.tsx
@@ -1,13 +1,15 @@
 import Button from '@Components/Button';
 import Icon from '@Components/Icon';
 import Input from '@Components/Input';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import styled from 'styled-components';
 
 export default function Chat ({ user }) {
 
+    const containerRef = useRef(null);
+
     const [messages, setMessages] = useState([]);
-    const [unconfirmedMessages, setUnconfirmedMessages] = useState([]); // [{content, isMine}
+    const [unconfirmedMessages, setUnconfirmedMessages] = useState([]);
     const [newMessage, setNewMessage] = useState('');
 
     useEffect(() => {
@@ -15,6 +17,10 @@ export default function Chat ({ user }) {
             window.app.Services.Messages.setCallback(null, null);
         }
     }, []);
+
+    useEffect(() => {
+        if(containerRef.current) containerRef.current.scrollTop = containerRef.current.scrollHeight;
+    }, [messages]);
 
     useEffect(() => {
         if(user) {
@@ -31,7 +37,7 @@ export default function Chat ({ user }) {
                             idSeen[message.id] = true;
                         }
                         messages.sort((a, b) => (new Date(a.date) as any) - (new Date(b.date) as any));
-                        return messages; 
+                        return messages;
                     });
                     setUnconfirmedMessages([]);
                 });
@@ -42,9 +48,19 @@ export default function Chat ({ user }) {
         }
     }, [user]);
 
+    const handleKeyDown = (e) => {
+        if(e.key === 'Enter') sendMessage();
+    }
+
+    const sendMessage = () => {
+        window.app.Services.Messages.sendPrivateMessage(user, newMessage);
+        setNewMessage('');
+        setUnconfirmedMessages((prev) => [...prev, { content: newMessage, isMine: true }]);
+    }
+
     return (
         <>
-            <MessagesContainer>
+            <MessagesContainer ref={containerRef}>
                 {(messages || [])?.map((message, i) => (
                     <Message
                         key={i}
@@ -71,13 +87,12 @@ export default function Chat ({ user }) {
                     value={newMessage}
                     onChange={(e) => setNewMessage(e.target.value)}
                     placeholder="Message"
+                    onKeyDown={handleKeyDown}
+                    className="focusable"
                 />
                 <Button
-                    onClick={() => {
-                        window.app.Services.Messages.sendPrivateMessage(user, newMessage);
-                        setNewMessage('');
-                        setUnconfirmedMessages((prev) => [...prev, { content: newMessage, isMine: true }]);
-                    }}
+                    className="focusable"
+                    onClick={sendMessage}
                 >
                     <Icon name="arrow-right"/>
                 </Button>
@@ -90,12 +105,12 @@ export default function Chat ({ user }) {
 
 const SendMessage = styled.div`
     display: flex;
-    gap: 10px;
-    padding: 10px;
+    gap: var(--quintet);
+    padding: var(--quintet);
     box-sizing: border-box;
-
+    border-top: var(--unit) solid rgba(255, 255, 255, 0.1);
     & > *:last-child {
-        width: 50px;
+        width: calc(var(--decade) * 4);
     }
 `;
 
@@ -103,19 +118,23 @@ const MessagesContainer = styled.div`
     height: 100%;
     width: 100%;
     overflow-y: auto;
-    padding: 10px;
+    padding: var(--quintet);
     display: flex;
     flex-direction: column;
-    gap: 10px;
+    gap: var(--quintet);
     box-sizing: border-box;
 `;
 
 const Message = styled.div`
-    padding: 10px;
-    border-radius: ${(props: any) => props.$isMine ? '10px 0px 10px 10px' : '0px 10px 10px 10px'};
+    padding: var(--quintet) var(--decade);
+    border-radius: ${(props: any) => props.$isMine ? 'var(--small-radius) 0 var(--small-radius) var(--small-radius)' : '0 var(--small-radius) var(--small-radius) var(--small-radius)'};
     background: ${(props: any) => props.$isMine ? 'var(--main)' : 'rgba(255, 255, 255, 0.1)'};
     color: #fff;
     align-self: ${(props: any) => props.$isMine ? 'flex-end' : 'flex-start'};
     max-width: 80%;
     word-break: break-word;
+
+    & > div {
+        font-size: var(--font-size);
+    }
 `;
