+++ app-front/Components/HeadBar/index.tsx
@@ -12,42 +12,47 @@ export default function HeadBar({ children, drop = "revert"}) {
 }
 
 const Bar = styled.div`
-    margin-top: 50px;
     /* position: relative; */
-    width: 100%;
-    height: 87px;
+    width: calc(100%);
+    height: calc(var(--decade) * 5.5);
     background-color: #fff;
     display: flex;
     justify-content: start;
     align-items: center;
-    padding: 20px 50px;
-    gap: 50px;
+    padding: calc(var(--quintet) * 2.5)  calc(var(--decade) * 3.5);
+    gap: calc(var(--decade) * 2);
     box-sizing: border-box;
-    border-radius: 28px;
+    border-radius: var(--radius);
     background: rgba(255, 255, 255, 0.10);
     z-index: 1;
     overflow: hidden;
+
+    &::-webkit-scrollbar {
+        display: none;
+    }
 
     backdrop-filter: ${(props: any) => props.$drop};
     
     & > div{
         position: relative;
         cursor: pointer;
-        font-size: 20px;
+        font-size: calc(var(--quintet) * 2.5) ;
         font-weight: 700;
         opacity: 0.5;
         color: #fff;
+        padding: calc(var(--unit) * 1) calc(var(--decade) * 1.5 / 2);
+        border-radius: var(--small-radius);
     }
 
     & > div:not(:last-child)::after{
         content: '';
         position: absolute;
         opacity: 0.1;
-        right: -25px;
+        right: calc(var(--quintet) * -3);
         top: 0;
         height: 100%;
-        width: 2px;
-        border-radius: 2px;
+        width: var(--unit);
+        border-radius: var(--unit);
         background: linear-gradient(180deg, rgba(255, 255, 255, 0.00) 0.79%, #FFF 57.69%, rgba(255, 255, 255, 0.00) 99.21%);
     }
 `;
