--- app-front/Components/SearchBar/index.tsx
@@ -1,49 +1,0 @@
-import { styled } from "styled-components";
-
-const SearchDiv = styled.div`
-    flex: 1 0 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 14px;
-    background: rgba(255, 255, 255, 0.1);
-    height: 65px;
-    padding: 15px 31px;
-    box-sizing: border-box;
-`;
-
-const SearchDivIcon = styled.div`
-    width: 30px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-`;
-
-const InputSearch = styled.input`
-    width: 100%;
-    font-family: "Poppins-Bold";
-    font-size: 20px;
-    color: white;
-`;
-
-export default function SearchBar() {
-    return (
-        <SearchDiv>
-            <SearchDivIcon>
-                <svg
-                    width="31"
-                    height="31"
-                    viewBox="0 0 31 31"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                >
-                    <path
-                        d="M13.625 0.166657C16.1519 0.166487 18.6279 0.877742 20.7695 2.21902C22.911 3.5603 24.6318 5.47742 25.7347 7.75097C26.8376 10.0245 27.2781 12.5627 27.0059 15.0749C26.7336 17.5871 25.7596 19.972 24.1953 21.9565L29.9776 27.7388C30.2661 28.0374 30.4256 28.4374 30.422 28.8525C30.4184 29.2677 30.2519 29.6648 29.9584 29.9584C29.6648 30.2519 29.2677 30.4184 28.8525 30.4221C28.4374 30.4257 28.0374 30.2661 27.7388 29.9777L21.9565 24.1953C20.2709 25.5237 18.2922 26.4297 16.1853 26.8378C14.0783 27.2458 11.9044 27.1442 9.84479 26.5412C7.78517 25.9382 5.89962 24.8515 4.34543 23.3716C2.79125 21.8917 1.6135 20.0615 0.910456 18.0339C0.207411 16.0063 -0.000548533 13.8399 0.303929 11.7156C0.608406 9.59121 1.41649 7.57047 2.66076 5.82193C3.90504 4.0734 5.54942 2.64778 7.45671 1.66401C9.36401 0.680233 11.4789 0.166839 13.625 0.166657ZM13.625 3.33332C10.8955 3.33332 8.27773 4.41762 6.34767 6.34768C4.4176 8.27774 3.33331 10.8955 3.33331 13.625C3.33331 16.3545 4.4176 18.9722 6.34767 20.9023C8.27773 22.8324 10.8955 23.9167 13.625 23.9167C16.3545 23.9167 18.9722 22.8324 20.9023 20.9023C22.8323 18.9722 23.9166 16.3545 23.9166 13.625C23.9166 10.8955 22.8323 8.27774 20.9023 6.34768C18.9722 4.41762 16.3545 3.33332 13.625 3.33332ZM13.625 4.91666C15.9346 4.91666 18.1496 5.83414 19.7827 7.46727C21.4158 9.1004 22.3333 11.3154 22.3333 13.625C22.3333 15.9346 21.4158 18.1496 19.7827 19.7827C18.1496 21.4158 15.9346 22.3333 13.625 22.3333C11.3154 22.3333 9.10038 21.4158 7.46725 19.7827C5.83412 18.1496 4.91664 15.9346 4.91664 13.625C4.91664 11.3154 5.83412 9.1004 7.46725 7.46727C9.10038 5.83414 11.3154 4.91666 13.625 4.91666Z"
-                        fill="white"
-                    />
-                </svg>
-            </SearchDivIcon>
-            <InputSearch type={"text"} placeholder={"Search for a game"} />
-        </SearchDiv>
-    );
-}
