+++ app-front/Components/WindowControls/index.tsx
@@ -55,19 +55,19 @@ export function FullscreenControls() {
 
 const FullscreenControlsContainer = styled.div`
     transition-duration: revert;
-    height: 30px;
-    width: 30px;
+    height: var(--window-handle);
+    width: var(--window-handle);
     display: flex;
     justify-content: center;
     align-items: center;
     position: absolute;
     cursor: pointer;
-    padding: 6px;
+    padding: calc(var(--unit) * 4);
     box-sizing: border-box;
     top: 0;
     right: 0;
     z-index: 10000;
-    border-radius: 0 0 0 8px;
+    border-radius: 0 0 0 var(--mini-radius);
     &:hover {
         background-color: rgba(255, 255, 255, 0.1);
     }
@@ -75,11 +75,11 @@ const FullscreenControlsContainer = styled.div`
 
 const ControlsContainer = styled.div`
     transition-duration: revert;
-    height: 30px;
+    height: var(--window-handle);
     display: flex;
     background-color: rgba(255, 255, 255, 0.05);
     background-color: rgba(255, 255, 255, 0.05);
-    border: 1px solid rgba(255, 255, 255, 0.1);
+    border: var(--border-size) solid rgba(255, 255, 255, 0.1);
     position: relative;
     z-index: 10000;
 
@@ -89,7 +89,7 @@ const ControlsContainer = styled.div`
         left: 0;
         top: 0;
         width: 100%;
-        height: 1px;
+        height: var(--unit)
     }
 
     & > div:first-child {
@@ -98,21 +98,22 @@ const ControlsContainer = styled.div`
         height: 100%;
         display: flex;
         align-items: center;
-        padding: 0 10px;
+        padding: 0 6px;
         font-weight: 600;
+        font-size: 12px;
     }
 
     & > div:last-child {
         display: flex;
 
         & > div {
-            width: 30px;
-            height: 30px;
+            width: var(--window-handle);
+            height: var(--window-handle);
             cursor: pointer;
             display: flex;
             align-items: center;
             justify-content: center;
-            padding: 6px;
+            padding: 4px;
             box-sizing: border-box;
             background-color: rgba(255, 255, 255, 0.05);
