+++ app-front/index.css
@@ -1,5 +1,17 @@
 :root {
-    --padding: 50px;
+    --scale: 1;
+    --unit: calc(0.125vh * var(--scale));
+    --quintet: calc(var(--unit) * 5);
+    --decade: calc(1.25vh * var(--scale));
+    --mini-radius: calc(var(--unit) * 5 * var(--scale));
+    --small-radius: calc(var(--unit) * 10 * var(--scale));
+    --radius: calc(var(--unit) * 16 * var(--scale));
+    --border-size: calc(0.125vh * var(--scale));
+    --padding: calc(var(--decade) * 3);
+    --nav-height: calc(var(--decade) * 6);
+    --nav-gap: calc(var(--decade) * 4);
+    --font-size: calc(var(--quintet) * 2.5);
+    --window-handle: 20px;
 }
 
 html,
@@ -33,8 +45,8 @@ body {
 }
 
 .focused {
-    outline: 5px solid #fff;
-    outline-offset: 1px;
+    outline: calc(var(--unit) * 4) solid #fff;
+    outline-offset: var(--unit);
 }
 
 *:not(.selectable) {
@@ -51,17 +63,6 @@ body {
 
 *::-webkit-scrollbar-track {
     background-color: transparent;
-}
-
-*::-webkit-scrollbar-track-piece {
-    margin: 20px 10px;
-    width: 10px;
-}
-
-*::-webkit-scrollbar-thumb {
-    background-color: #00000040;
-    border-radius: 24px;
-    width: 10px;
 }
 
 body {
@@ -236,13 +237,4 @@ a {
     font-family: "Montserrat-ThinItalic";
     src: url("./../fonts/Montserrat-ThinItalic.ttf") format("truetype");
 }
-._an_r {
-    position: absolute;
-    left: 0;
-    width: 0;
-    height: 50px;
-    background: white;
-    border-radius: 50px;
-    opacity: 1;
-    transition: 0.25s ease-out;
-}
+
